Lock input but don't lock screen
This patch keeps the screen unlocked but keeps the input locked.
That is, the screen is not affected by slock, but users will not
be able to interact with the X session unless they enter the correct password.

There's a lot of extraneous code that can be removed since the goal is
for input to be locked while at the same time allowing the screen to be
updated and viewable. But this patch just works (tm) and I don't want to
potentially introduce bugs.

Signed-off-by: drozdowsky <hdrozdow+github@pm.me>